requirements.txt
aiohttp>=3.8.3
aiogram==3.2.0
requests~=2.26.0
python-dotenv==1.0.0
